requirements.txt
yt-dlp>=2024.1.1
openai-whisper>=20231117
google-generativeai>=0.5.0
python-dotenv>=1.0.0
rich>=13.7.0
ffmpeg-python>=0.2.0
